+++ packages/semver/src/generators/install/schema.json
@@ -1,7 +1,8 @@
 {
   "$schema": "http://json-schema.org/schema",
   "$id": "ng-add",
-  "title": "@jscutlery/semver ng-add schematic",
+  "cli": "nx",
+  "title": "@jscutlery/semver install generator",
   "type": "object",
   "properties": {
     "syncVersions": {
